+++ errmap/pkg/errmap/errmap_test.go
@@ -3,6 +3,7 @@ package errmap_test
 import (
 	"errors"
 	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/puppetlabs/leg/errmap/pkg/errmap"
@@ -53,4 +54,44 @@ func TestMapApply(t *testing.T) {
 
 	assert.True(t, errors.Is(err, cause))
 	assert.EqualError(t, err, "c: OM NOM NOM: b: a: foo")
+}
+
+type joinError struct {
+	bases []error
+}
+
+func (e *joinError) Error() string {
+	ps := make([]string, len(e.bases))
+	for i, p := range e.bases {
+		ps[i] = p.Error()
+	}
+	return fmt.Sprintf("joined %d error(s): %s", len(e.bases), strings.Join(ps, ", "))
+}
+
+func (e *joinError) MapApply(m errmap.Mapper) error {
+	ne := &joinError{}
+	for _, base := range e.bases {
+		next := m.Map(base)
+
+		switch nt := next.(type) {
+		case *joinError:
+			ne.bases = append(ne.bases, nt.bases...)
+		default:
+			ne.bases = append(ne.bases, next)
+		}
+	}
+	return ne
+}
+
+func TestMapApplyRecursive(t *testing.T) {
+	err := errmap.Wrap(errors.New("foo"), "first")
+	err = errmap.MapLast(err, errmap.MapperFunc(func(err error) error {
+		inner := errmap.MapLast(err, errmap.MapperFunc(func(err error) error {
+			return &joinError{bases: []error{fmt.Errorf("inner: %w", err)}}
+		}))
+		return &joinError{bases: []error{inner, err}}
+	}))
+	err = errmap.Wrap(err, "outer")
+
+	assert.EqualError(t, err, "joined 2 error(s): outer: inner: first: foo, outer: first: foo")
 }